pyproject.toml
# SPDX-License-Identifier: MIT
#
# SPDX-Copyright-Text: © 2017 PySSB contributors (see AUTHORS for more details)
[tool.poetry]
name = "ssb"
version = "0.1.0"
description = "Secure Scuttlebutt library in Python"
authors = ["PyScuttleButt Contributors <pedro@dete.st>"]
license = "MIT"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.9"
async_generator = "^1.8"
PyNaCl = "^1.1.0"
PyYAML = "^6.0.1"
secret-handshake = { version = "0.1.0.dev3", allow-prereleases = true }
simplejson = "3.16.0"
colorlog = "^6.7.0"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
check-manifest = "^0.39"
commitizen = "^3.12.0"
coverage = "^7.3.2"
isort = "^4.3.20"
pep257 = "^0.7.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"

[tool.poetry.group.docs.dependencies]
Sphinx = "^2.1.1"

[tool.black]
line-length = 120

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 70
omit = ["examples/*"]

[tool.pytest.ini_options]
addopts = ["--cov=.", "--no-cov-on-fail"]
python_files = ["tests/test_*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
